requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.2
fonttools==4.49.0
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.5
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.4
openslide-python==1.3.1
packaging==24.0
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
shapely==2.0.3
six==1.16.0
threadpoolctl==3.3.0
tifffile==2024.2.12
Werkzeug==3.0.1
